requirements.txt
fastapi>=0.110,<1
gunicorn>=21,<24
a2wsgi>=1.7,<2
opentelemetry-proto>=1.20,<2
protobuf>=4.21,<6
